FilteringTerm entity
The current definition just provides the multiple for `filteringTerms`, as an array populated by the selected types (i.e. OntologyFilter, AlphanumericFilter, CustomFilter). However, (AFAIK) the proper way would be to define a `FilteringTerm` prototype which can be one of those instances; otherwise there is no `FilteringTerm` as a concept.
This PR introduces the `FilteringTerm` object as a wrapper for individual instances.