Defining structures for multiple Boolean/Count responses ...
This is a first pass to introduce the option of multiple Boolean or count responses, similar to resultSets, to allow aggregation of responses e.g. in network settings w/o definition of a new message type.

This proposal is up for editing (e.g. is `responseSets` as new parameter o.k. or should it be completely aligned w/ the resultSets respose?).