first edit of requestParameterComponents
This commit is the first, incomplete, edit for `requestParameterComponents`. This is the result of the Connect 2025 discussions w/ the agreement that parameters should be defined globally in the model and then referenced from whatever endpoint definitions are provided (e.g. OpenAPI) or generated.

This commit still lacks the list of request parameters itself as well as the requestProfiles (another Connect discussion outcome).

minor path rewrite fix