removing CountEntity from beaconAggregateResults
Since this aggregate ... implementation uses standard Beacon requests and responses, counts should always reflect the entry type of the response; i.e. it should be the count of the `individuals` having `biosamples` with a certain histology if using `.../individuals/?...`. While certainly _possible_ it seems pretty confusing to have different entities counted. Also it is not intuitive since the response entity indicates _what_ should be reported on.